Initial dev commit
This is the initial incomplete proposal of wikibase-docker-yml-gui so know as blubber. It is dealing with the following issues:
	1. separate program flow from the datastructure including the template-based program flow 	of vuejs and avoids this concept of vuejs;
	2. generate from a given configuration dynamicly the form;
	3. seperate the displayed strings from markup, respectivly avoid hardcoded strings.

To the first and second point - this draft is mainly implemented via Vue.mixin. Via mixin it is possible to provide a cleaner encapsulation instead of implement it all in the main.js. Also vuejs will not use render functionality in a component and mixin can solve that. The BlubblerForm.vue is supposed to be the main location to do that as well as the place to load the configuration (src/components/data/config.json). To create the form this application will use the packages „vue-form-wizard“ and „vue-form-generator“. BlubblerForm.vue will generate the schema, options and model for form-generator. The form-generator will create the tempate for form-wizard and the form-wizard at least will build the form.

To the third point – Language.vue will provide this functionality as well as the basic concepts vuejs. So far Language.vue exists only as a skeleton, but is supposed to use i18n in the futher process. It should also load the necessary string from src/components/lang/<language-code>.json

Utils.vue is supposed to be a auxilary component and provides at this point a import functionality for interal and external resources (get()). For that are tow things are needed – a string with the file, which should requested and a callback function. The callback is necessary to handle the asynchronous request. Also all internal resources must be located inside the components folder. For external request Utils.vue uses axios.

App.vue is so far responsible for importing the other components, but is supposed to handle the all futher form activities, if it‘s possible.