Futher Process
This commit introduce a clearer seperation of the programmflow. The render is now binded at the App.vue (buildApplication). It took a bit to figure out, that vue-js need the template property at a component setted to use the render functionality. Also do not use render at mixin and the template property at a component in the same time – this will cause a infinite loop. The BlubberFormBuilder is also now a little bit more abstract and does not depend on a inputbinding of the vue object itself.

The Language.vue is now capible to load a language file and create dynamicly a i18n object.

The YMLBuilder is supposed to create the yml-output in the futher work.

The next step will be to complete the formBuilder.