Futher process
The BlubberFormFactory is now capable of dealing with the common properties of vue-form-generator. Also the order of methods was restructured to keep clean structure. AssignValueOrFunction has been removed.

There also some ajustments of Utils - it got a very simple
implementation of substring search.

Next will be to ajust vue-generator element and so the form should show
up.