Refactoring
The code is now divied in smaller chunks. In case of utils the code is splitted also in several files.
Language and BlubberFormFactory are not longer global mixinis, but local
onces. Utils is now a class with static methodes, and the prototypes are
removed and reimplemented in the different helpers.

Also there are 2 small improvements - StringHelper.format is now able
to take objects and arrays as 2nd parameter. The keys are the
placeholder in the formatted strings.
Ultis has now a waitUntil method, which takes a callback function to
ensure, that a specific async value is loaded.